Initial Commit
Adding file that pip installs virtualenv.

Adding file that creates a virtualenv in the directory where file lives. Name is "default".

The idea is a user who doesn't want to use a commandline uses an IDE of their choice to install virtualenv with pip_install_virtualenv.py and then creates a virtualenv with create_virtualenv.py.